+++ src/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <title><%= htmlWebpackPlugin.options.title %></title>
-    <script src="preferences.js"></script>
+    <script src="runtime-config.js"></script>
   </head>
   <body>
   </body>
